osa4 tehtävät

+++ osa4/blogilista/controllers/blogs.js
@@ -1,32 +1,58 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
   const body = request.body
 
   if (!body.title || !body.url) {
-    return response.status(400).json({error: 'title or url missing'})
+    return response.status(400).json({ error: 'title or url missing' })
   }
 
+  const user = request.user
+
   const blog = body.likes ?
-  new Blog(body) :
   new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: 0
+    likes: body.likes,
+    user: user._id
+  }) :
+  new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: 0,
+    user: user._id
   })
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+  
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
+
+  if (!blog) {
+    return response.status(204).end()
+  }
+
+  if (blog.user.toString() !== user.id.toString()) {
+    return response.status(401).json({ error: 'no permission to delete this resource' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
